+++ src/customer/PaymentProducts.java
@@ -127,7 +127,7 @@ public class PaymentProducts implements KeyListener {
         pro_payment_btn_confirm.addActionListener(new ActionListener() {
             @Override
             public void actionPerformed(ActionEvent e) {
-                if(Integer.parseInt(pro_payment_amount.getText())<=Integer.parseInt(CustomerProfileData.getBankBalance())){
+                if(Float.parseFloat(pro_payment_amount.getText())<=Float.parseFloat(CustomerProfileData.getBankBalance())){
                 if (getAuthentication()) {
                     WalletDataG.setBalence(Float.valueOf(CustomerProfileData.getBankBalance()));//customer Balence
                     WalletDataG.setTransationType("Products");
